fix 64 to 32 bit clang conversion warning in src/core/lib

+++ src/cpp/client/channel_test_peer.cc
@@ -24,7 +24,7 @@
 namespace grpc {
 namespace testing {
 
-int ChannelTestPeer::registered_calls() const {
+size_t ChannelTestPeer::registered_calls() const {
   return grpc_core::Channel::FromC(channel_->c_channel_)
       ->TestOnlyRegisteredCalls();
 }
